requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.10
coreapi==2.3.3
coreschema==0.0.4
dj-config-url==0.1.1
dj-database-url==0.5.0
Django==4.0.1
django-cors-headers==3.11.0
django-environ==0.8.1
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
psycopg2==2.9.3
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
sqlparse==0.4.2
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.8
whitenoise==5.3.0
